Move `Route` finding functions to `RouteFinder.cpp`
