+++ appOPHD/appOPHD.vcxproj.filters
@@ -87,6 +87,9 @@
     <ClCompile Include="Map\MapView.cpp">
       <Filter>Source Files\Map</Filter>
     </ClCompile>
+    <ClCompile Include="Map\RouteFinder.cpp">
+      <Filter>Source Files\Map</Filter>
+    </ClCompile>
     <ClCompile Include="Map\Tile.cpp">
       <Filter>Source Files\Map</Filter>
     </ClCompile>
@@ -390,6 +393,9 @@
       <Filter>Header Files\Map</Filter>
     </ClInclude>
     <ClInclude Include="Map\MapView.h">
+      <Filter>Header Files\Map</Filter>
+    </ClInclude>
+    <ClInclude Include="Map\RouteFinder.h">
       <Filter>Header Files\Map</Filter>
     </ClInclude>
     <ClInclude Include="Map\Tile.h">
